moved extract_features to seperate module, baseline.py now loads from file
